debug

+++ ft_substr.c
@@ -18,9 +18,9 @@ char	*ft_substr(char const *s, unsigned int start, size_t len)
 
 	if (!s)
 		return (0);
-	if (ft_strlen(s) < start)
+	if ((unsigned int) ft_strlen(s) < start)
 		len = 0;
-	if (ft_strlen(s + start) < len)
+	if ((unsigned int) ft_strlen(s + start) < len)
 		len = ft_strlen(s + start);
 	str = malloc (sizeof(char) * (len + 1));
 	if (!str)
